SNMP protocol version is now set in preferences

+++ src/NetworkInterface.cpp
@@ -4665,7 +4665,8 @@ void NetworkInterface::getnDPIProtocols(lua_State *vm, ndpi_protocol_category_t 
   for(i=0; i<(int)ndpi_struct->ndpi_num_supported_protocols; i++) {
     char buf[8];
 
-    if(((filter == (u_int8_t)-1) || ndpi_struct->proto_defaults[i].protoCategory == filter) &&
+    if((((u_int8_t)filter == (u_int8_t)-1)
+	|| ndpi_struct->proto_defaults[i].protoCategory == filter) &&
 	(!skip_critical || !Utils::isCriticalNetworkProtocol(i))) {
       snprintf(buf, sizeof(buf), "%d", i);
       lua_push_str_table_entry(vm, ndpi_struct->proto_defaults[i].protoName, buf);
